+++ VWOSdk.Tests/Mock/MockCampaignResolver.cs
@@ -27,13 +27,19 @@ namespace VWOSdk.Tests
             return new Mock<ICampaignAllocator>();
         }
 
-        internal static void SetupResolve(Mock<ICampaignAllocator> mockCampaignResolver, BucketedCampaign allocateCampaignReturnValue, BucketedCampaign getCampaignReturnValue = null)
+        internal static void SetupResolve(Mock<ICampaignAllocator> mockCampaignResolver, BucketedCampaign allocateCampaignReturnValue, BucketedCampaign getCampaignReturnValue = null, BucketedCampaign otherAllocateCampaignReturnValue = null)
         {
             mockCampaignResolver.Setup(mock => mock.Allocate(It.IsAny<AccountSettings>(), It.IsAny<UserStorageMap>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>()))
                 .Returns(allocateCampaignReturnValue);
 
+            mockCampaignResolver.Setup(mock => mock.Allocate(It.IsAny<AccountSettings>(), It.IsAny<UserStorageMap>(), "MockCampaignKey1", It.IsAny<string>(), It.IsAny<string>()))
+                .Returns(otherAllocateCampaignReturnValue);
+
             mockCampaignResolver.Setup(mock => mock.GetCampaign(It.IsAny<AccountSettings>(), It.IsAny<string>()))
                 .Returns(getCampaignReturnValue);
+
+            mockCampaignResolver.Setup(mock => mock.GetCampaign(It.IsAny<AccountSettings>(), "MockCampaignKey1"))
+                .Returns(otherAllocateCampaignReturnValue);
         }
     }
 }