+++ VWOSdk/Internal/Validator/Validator.cs
@@ -63,6 +63,8 @@ namespace VWOSdk
             result = ValidateWithLog(() => ValidateString(userId), nameof(userId), nameof(Track)) && result;
             result = ValidateWithLog(() => ValidateString(goalIdentifier), nameof(goalIdentifier), nameof(Track)) && result;
             var customVariables = options.ContainsKey("customVariables") ? options["customVariables"] : null;
+            String goalTypeToTrack = options.ContainsKey("goalTypeToTrack") ? options["goalTypeToTrack"] : null;
+            result = ValidateWithLog(() => goalTypeToTrack == null || Constants.GoalTypes.VALUES.Contains(goalTypeToTrack), nameof(goalIdentifier), nameof(Track)) && result;
             return ValidateWithLog(() => ValidateNullableFloat(revenueValue) && (customVariables == null || customVariables is Dictionary<string, dynamic>) , nameof(revenueValue), nameof(Track)) && result;
         }
 
@@ -88,7 +90,7 @@ namespace VWOSdk
         public bool SettingsFile(Settings settingsFile)
         {
             var result = NotNull(settingsFile);
-            result = result && ValidateLong(settingsFile.AccountId) && ValidateString(settingsFile.SdkKey);
+            result = result && ValidateLong(settingsFile.AccountId);
             result = result && Validate(settingsFile.Campaigns);
             return result;
         }